# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediConnect - Patient and Doctor Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * @nullable
 */
export type UserProfileRole =
  | (typeof UserProfileRole)[keyof typeof UserProfileRole]
  | null;

export const UserProfileRole = {
  patient: "patient",
  doctor: "doctor",
} as const;

export interface UserProfile {
  id: number;
  clerkId: string;
  /** @nullable */
  role?: UserProfileRole;
  name: string;
  email: string;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  heightCm?: number | null;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  conditions?: string | null;
  /** @nullable */
  qualifications?: string | null;
  /** @nullable */
  specialty?: string | null;
  createdAt: string;
}

export interface UpdateUserProfileBody {
  name?: string;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  heightCm?: number | null;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  conditions?: string | null;
  /** @nullable */
  qualifications?: string | null;
  /** @nullable */
  specialty?: string | null;
}

export type CompleteOnboardingBodyRole =
  (typeof CompleteOnboardingBodyRole)[keyof typeof CompleteOnboardingBodyRole];

export const CompleteOnboardingBodyRole = {
  patient: "patient",
  doctor: "doctor",
} as const;

export interface CompleteOnboardingBody {
  role: CompleteOnboardingBodyRole;
  name: string;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  heightCm?: number | null;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  conditions?: string | null;
  /** @nullable */
  qualifications?: string | null;
  /** @nullable */
  specialty?: string | null;
}

export interface PatientProfile {
  id: number;
  clerkId: string;
  name: string;
  email: string;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  heightCm?: number | null;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  conditions?: string | null;
}

export interface DoctorProfile {
  id: number;
  clerkId: string;
  name: string;
  email: string;
  /** @nullable */
  qualifications?: string | null;
  /** @nullable */
  specialty?: string | null;
  /** @nullable */
  nextAppointment?: string | null;
}

export interface AddPatientBody {
  patientClerkId: string;
}

export interface Conversation {
  contactId: string;
  contactName: string;
  contactRole: string;
  /** @nullable */
  lastMessage?: string | null;
  /** @nullable */
  lastMessageAt?: string | null;
  unreadCount: number;
}

export interface Message {
  id: number;
  senderId: string;
  receiverId: string;
  content: string;
  sentAt: string;
  isRead: boolean;
}

export interface SendMessageBody {
  content: string;
}

export interface AppointmentSlot {
  id: number;
  doctorId: string;
  doctorName: string;
  slotDate: string;
  slotTime: string;
  isAvailable: boolean;
  createdAt: string;
}

export interface AddAppointmentSlotBody {
  slotDate: string;
  slotTime: string;
}

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  pending: "pending",
  accepted: "accepted",
  declined: "declined",
} as const;

export interface Appointment {
  id: number;
  patientId: string;
  patientName: string;
  doctorId: string;
  doctorName: string;
  requestedDate: string;
  requestedTime: string;
  status: AppointmentStatus;
  /** @nullable */
  doctorNote?: string | null;
  createdAt: string;
}

export interface RequestAppointmentBody {
  requestedDate: string;
  requestedTime: string;
}

export type RespondToAppointmentBodyStatus =
  (typeof RespondToAppointmentBodyStatus)[keyof typeof RespondToAppointmentBodyStatus];

export const RespondToAppointmentBodyStatus = {
  accepted: "accepted",
  declined: "declined",
} as const;

export interface RespondToAppointmentBody {
  status: RespondToAppointmentBodyStatus;
  /** @nullable */
  doctorNote?: string | null;
}

export interface VitalLog {
  id: number;
  patientId: string;
  /** @nullable */
  heartRate?: number | null;
  /** @nullable */
  respirationRate?: number | null;
  /** @nullable */
  systolicBp?: number | null;
  /** @nullable */
  diastolicBp?: number | null;
  /** @nullable */
  spo2?: number | null;
  isCritical: boolean;
  isDismissed: boolean;
  loggedAt: string;
  /** @nullable */
  notes?: string | null;
}

export interface AddVitalLogBody {
  /** @nullable */
  heartRate?: number | null;
  /** @nullable */
  respirationRate?: number | null;
  /** @nullable */
  systolicBp?: number | null;
  /** @nullable */
  diastolicBp?: number | null;
  /** @nullable */
  spo2?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface CriticalVital {
  logId: number;
  patientId: string;
  patientName: string;
  /** @nullable */
  heartRate?: number | null;
  /** @nullable */
  respirationRate?: number | null;
  /** @nullable */
  systolicBp?: number | null;
  /** @nullable */
  diastolicBp?: number | null;
  /** @nullable */
  spo2?: number | null;
  isDismissed: boolean;
  loggedAt: string;
}

export interface Medication {
  id: number;
  patientId: string;
  doctorId: string;
  doctorName: string;
  medicationName: string;
  /** @nullable */
  dosage?: string | null;
  /** @nullable */
  frequency?: string | null;
  /** @nullable */
  timeToTake?: string | null;
  /** @nullable */
  conditionInfo?: string | null;
  /** @nullable */
  additionalInfo?: string | null;
  isCurrent: boolean;
  prescribedAt: string;
}

export interface MedicationList {
  current: Medication[];
  past: Medication[];
}

export interface AddMedicationBody {
  patientId: string;
  medicationName: string;
  /** @nullable */
  dosage?: string | null;
  /** @nullable */
  frequency?: string | null;
  /** @nullable */
  timeToTake?: string | null;
  /** @nullable */
  conditionInfo?: string | null;
  /** @nullable */
  additionalInfo?: string | null;
}

export interface UpdateMedicationBody {
  isCurrent?: boolean;
  /** @nullable */
  dosage?: string | null;
  /** @nullable */
  frequency?: string | null;
  /** @nullable */
  timeToTake?: string | null;
  /** @nullable */
  conditionInfo?: string | null;
  /** @nullable */
  additionalInfo?: string | null;
}

export interface PatientDashboardSummary {
  profile: UserProfile;
  doctor?: DoctorProfile;
  upcomingAppointments: Appointment[];
  topConversations: Conversation[];
  latestVitals?: VitalLog;
  upcomingMedications: Medication[];
}

export interface DoctorDashboardSummary {
  profile: UserProfile;
  upcomingAppointments: Appointment[];
  topConversations: Conversation[];
  criticalVitals: CriticalVital[];
  recentMedications: Medication[];
  pendingRequests: number;
}