# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MediConnect - Patient and Doctor Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments/slots:
    post:
      operationId: addAppointmentSlot
      tags: [appointments]
      summary: Doctor adds an available appointment slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddAppointmentSlotBody"
      responses:
        "201":
          description: Slot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentSlot"
        "400":
          description: Slot already exists or invalid
        "403":
          description: Forbidden - not a doctor
  /appointments/request:
    post:
      operationId: requestAppointment
      tags: [appointments]
      summary: Patient requests an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestAppointmentBody"
      responses:
        "201":
          description: Appointment request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "403":
          description: Forbidden - not a patient
  /appointments/requests:
    get:
      operationId: getAppointmentRequests
      tags: [appointments]
      summary: Get pending appointment requests for the doctor
      responses:
        "200":
          description: List of pending requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
        "403":
          description: Forbidden - not a doctor
  /appointments/{id}/respond:
    post:
      operationId: respondToAppointment
      tags: [appointments]
      summary: Doctor accepts or declines an appointment request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RespondToAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "403":
          description: Forbidden
        "404":
          description: Not found

  # Vitals Endpoints
...
  /medications/patient/{patientId}:
    get:
      operationId: getPatientMedications
      tags: [medications]
      summary: Get medications for a specific patient (doctor only)
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Patient medications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicationList"
        "403":
          description: Forbidden
  /medications:
    post:
      operationId: addMedication
      tags: [medications]
      summary: Doctor adds a medication for a patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMedicationBody"
      responses:
        "201":
          description: Medication created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
        "403":
          description: Forbidden - not a doctor
  /medications/{id}:
    patch:
      operationId: updateMedication
      tags: [medications]
      summary: Update a medication (e.g. mark as past)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMedicationBody"
      responses:
        "200":
          description: Updated medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
        "403":
          description: Forbidden
        "404":
          description: Not found

  # Dashboard/Summary Endpoints
...
components:
  schemas:
...
    AppointmentSlot:
      type: object
      properties:
        id:
          type: integer
        doctorId:
          type: string
        doctorName:
          type: string
        slotDate:
          type: string
        slotTime:
          type: string
        isAvailable:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - doctorId
        - doctorName
        - slotDate
        - slotTime
        - isAvailable
        - createdAt
    AddAppointmentSlotBody:
      type: object
      properties:
        slotDate:
          type: string
        slotTime:
          type: string
      required:
        - slotDate
        - slotTime
    Appointment:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: string
        patientName:
          type: string
        doctorId:
          type: string
        doctorName:
          type: string
        requestedDate:
          type: string
        requestedTime:
          type: string
        status:
          type: string
          enum: ["pending", "accepted", "declined"]
        doctorNote:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - patientId
        - patientName
        - doctorId
        - doctorName
        - requestedDate
        - requestedTime
        - status
        - createdAt
    RequestAppointmentBody:
      type: object
      properties:
        requestedDate:
          type: string
        requestedTime:
          type: string
      required:
        - requestedDate
        - requestedTime
    RespondToAppointmentBody:
      type: object
      properties:
        status:
          type: string
          enum: ["accepted", "declined"]
        doctorNote:
          type: ["string", "null"]
      required:
        - status
...
    Medication:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: string
        doctorId:
          type: string
        doctorName:
          type: string
        medicationName:
          type: string
        dosage:
          type: ["string", "null"]
        frequency:
          type: ["string", "null"]
        timeToTake:
          type: ["string", "null"]
        conditionInfo:
          type: ["string", "null"]
        additionalInfo:
          type: ["string", "null"]
        isCurrent:
          type: boolean
        prescribedAt:
          type: string
      required:
        - id
        - patientId
        - doctorId
        - doctorName
        - medicationName
        - isCurrent
        - prescribedAt
    MedicationList:
      type: object
      properties:
        current:
          type: array
          items:
            $ref: "#/components/schemas/Medication"
        past:
          type: array
          items:
            $ref: "#/components/schemas/Medication"
      required:
        - current
        - past
    AddMedicationBody:
      type: object
      properties:
        patientId:
          type: string
        medicationName:
          type: string
        dosage:
          type: ["string", "null"]
        frequency:
          type: ["string", "null"]
        timeToTake:
          type: ["string", "null"]
        conditionInfo:
          type: ["string", "null"]
        additionalInfo:
          type: ["string", "null"]
      required:
        - patientId
        - medicationName
    UpdateMedicationBody:
      type: object
      properties:
        isCurrent:
          type: boolean
        dosage:
          type: ["string", "null"]
        frequency:
          type: ["string", "null"]
        timeToTake:
          type: ["string", "null"]
        conditionInfo:
          type: ["string", "null"]
        additionalInfo:
          type: ["string", "null"]
...